+++ lib/features/login/data/mapper/auth_mapper.dart
@@ -1,7 +1,7 @@
 import 'package:good_app/app/core/mappers/mapper_entity.dart';
 import 'package:good_app/app/core/mappers/mapper_model.dart';
-import 'package:good_app/features/login/data/models/auth.dart';
-import 'package:good_app/features/login/domain/entities/auth.dart';
+import 'package:good_app/features/login/data/models/auth_model.dart';
+import 'package:good_app/features/login/domain/entities/auth_entity.dart';
 
 abstract class AuthMapper
     with
