+++ lib/features/login/domain/entities/auth_entity.dart
