+++ lib/features/login/domain/entities/user_entity.dart
