+++ lib/features/login/domain/repositories/login_repository.dart
@@ -1,7 +1,7 @@
 import 'package:dartz/dartz.dart';
 
 import 'package:good_app/app/core/errors/default/default_error.dart';
-import 'package:good_app/features/login/domain/entities/auth.dart';
+import 'package:good_app/features/login/domain/entities/auth_entity.dart';
 
 abstract class LoginRepository {
   Future<Either<DefaultError, AuthEntity>> call({
