+++ lib/features/login/domain/usecases/do_login.dart
@@ -2,7 +2,7 @@ import 'package:dartz/dartz.dart';
 
 import 'package:good_app/app/core/errors/default/default_error.dart';
 import 'package:good_app/app/core/usecase/usecase.dart';
-import 'package:good_app/features/login/domain/entities/auth.dart';
+import 'package:good_app/features/login/domain/entities/auth_entity.dart';
 import 'package:good_app/features/login/domain/repositories/login_repository.dart';
 
 class LoginParams {
