+++ test/features/login/data/repositories/login_repository_impl_test.dart
@@ -1,6 +1,6 @@
 import 'package:dartz/dartz.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:good_app/features/login/data/models/auth.dart';
+import 'package:good_app/features/login/data/models/auth_model.dart';
 import 'package:mocktail/mocktail.dart';
 
 import 'package:good_app/app/core/errors/failures/password_wrong.dart';
